requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.9.14
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.3
hydra-core==1.2.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
threadpoolctl==3.1.0
torch==1.12.1
torchsummary==1.5.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.12
